+++ modules/manager/logging/encoding.go
@@ -30,8 +30,9 @@ var (
 // SetupLogEncoder sets additional logger configuration options when development mode is enabled.
 // In this way, the log structure is lighter and more human-friendly when the development mode
 // is enabled.
-func SetupLogEncoder(developmentMode bool, options *zap.Options) *zap.Options {
-	if developmentMode {
+func SetupLogEncoder(loggingMode Mode, options *zap.Options) *zap.Options {
+	if loggingMode == DevelopmentMode {
+		options.Development = true
 		options.TimeEncoder = defaultDevTimeEncoder
 		options.EncoderConfigOptions = []zap.EncoderConfigOption{
 			func(ec *zapcore.EncoderConfig) {