+++ pkg/utils/kubernetes/resources/zz_generated.kic_validatingwebhookconfig.go
@@ -17,7 +17,7 @@ import (
 // GenerateValidatingWebhookConfigurationForControlPlane generates a ValidatingWebhookConfiguration for a control plane
 // based on the control plane version. It also overrides all webhooks' client configurations with the provided service
 // details.
-func GenerateValidatingWebhookConfigurationForControlPlane(webhookName string, image string, devMode bool, clientConfig admregv1.WebhookClientConfig) (*admregv1.ValidatingWebhookConfiguration, error) {
+func GenerateValidatingWebhookConfigurationForControlPlane(webhookName string, image string, validateControlPlaneImage bool, clientConfig admregv1.WebhookClientConfig) (*admregv1.ValidatingWebhookConfiguration, error) {
 	if webhookName == "" {
 		return nil, fmt.Errorf("webhook name is required")
 	}
@@ -29,18 +29,18 @@ func GenerateValidatingWebhookConfigurationForControlPlane(webhookName string, i
 	// of the controlplane. When an invalid or unsupported image is used in dev mode,
 	// the clusterRole associated to the default ControlPlane image is used instead.
 	v, err := versions.FromImage(image)
-	if err != nil && !devMode {
+	if err != nil && validateControlPlaneImage {
 		return nil, err
 	}
 
 	supported, err := versions.IsControlPlaneImageVersionSupported(image)
-	if err != nil && !devMode {
+	if err != nil && validateControlPlaneImage {
 		return nil, err
 	}
-	if !devMode && !supported {
+	if validateControlPlaneImage && !supported {
 		return nil, ErrControlPlaneVersionNotSupported
 	}
-	if devMode && !supported {
+	if !validateControlPlaneImage && !supported {
 		v, err = semverv4.Parse(versions.DefaultControlPlaneVersion)
 		if err != nil {
 			return nil, fmt.Errorf("error when creating semver from the default controlplane version: %w", err)