+++ test/envtest/kongpluginbinding_unmanaged_test.go
@@ -15,6 +15,7 @@ import (
 
 	"github.com/kong/gateway-operator/controller/konnect"
 	sdkmocks "github.com/kong/gateway-operator/controller/konnect/ops/sdk/mocks"
+	"github.com/kong/gateway-operator/modules/manager/logging"
 	"github.com/kong/gateway-operator/modules/manager/scheme"
 	"github.com/kong/gateway-operator/test/helpers/deploy"
 
@@ -40,7 +41,7 @@ func TestKongPluginBindingUnmanaged(t *testing.T) {
 	sdk := factory.SDK
 
 	reconcilers := []Reconciler{
-		konnect.NewKonnectEntityReconciler(factory, false, mgr.GetClient(),
+		konnect.NewKonnectEntityReconciler(factory, logging.DevelopmentMode, mgr.GetClient(),
 			konnect.WithKonnectEntitySyncPeriod[configurationv1alpha1.KongPluginBinding](konnectInfiniteSyncTime),
 		),
 	}